Move JsonInclude tests for overrides into separate class
