+++ src/test/java/com/fasterxml/jackson/databind/ser/filter/JsonIncludeTest.java
@@ -149,40 +149,6 @@ public class JsonIncludeTest
         public NonDefaultCalendar(Calendar v) { super(v); }
     }
 
-    @JsonPropertyOrder({"list", "map"})
-    static class EmptyListMapBean
-    {
-        public List<String> list = Collections.emptyList();
-
-        public Map<String,String> map = Collections.emptyMap();
-    }
-
-    @JsonInclude(JsonInclude.Include.ALWAYS)
-    @JsonPropertyOrder({"num", "annotated", "plain"})
-    static class MixedTypeBean
-    {
-        @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
-        public Integer num = null;
-
-        @JsonInclude(JsonInclude.Include.NON_NULL)
-        public String annotated = null;
-
-        public String plain = null;
-    }
-
-    @JsonInclude(JsonInclude.Include.NON_NULL)
-    @JsonPropertyOrder({"num", "annotated", "plain"})
-    static class MixedTypeNonNullBean
-    {
-        @JsonInclude(JsonInclude.Include.USE_DEFAULTS)
-        public Integer num = null;
-
-        @JsonInclude(JsonInclude.Include.ALWAYS)
-        public String annotated = null;
-
-        public String plain = null;
-    }
-
     // [databind#1351]
 
     static class Issue1351Bean
@@ -338,142 +304,6 @@ public class JsonIncludeTest
         IntWrapper zero = new IntWrapper(0);
         assertEquals("{\"i\":0}", defMapper.writeValueAsString(zero));
         assertEquals("{\"i\":0}", inclMapper.writeValueAsString(zero));
-    }
-
-    public void testPropConfigOverridesForInclude() throws IOException
-    {
-        // First, with defaults, both included:
-        EmptyListMapBean empty = new EmptyListMapBean();
-        assertEquals(aposToQuotes("{'list':[],'map':{}}"),
-                MAPPER.writeValueAsString(empty));
-        ObjectMapper mapper;
-
-        // and then change inclusion criteria for either
-        mapper = new ObjectMapper();
-        mapper.configOverride(Map.class)
-            .setInclude(JsonInclude.Value.construct(JsonInclude.Include.NON_EMPTY, null));
-        assertEquals(aposToQuotes("{'list':[]}"),
-                mapper.writeValueAsString(empty));
-
-        mapper = new ObjectMapper();
-        mapper.configOverride(List.class)
-            .setInclude(JsonInclude.Value.construct(JsonInclude.Include.NON_EMPTY, null));
-        assertEquals(aposToQuotes("{'map':{}}"),
-                mapper.writeValueAsString(empty));
-    }
-
-    public void testPropConfigOverrideForIncludeAsPropertyNonNull() throws Exception
-    {
-        // First, with defaults, all but NON_NULL annotated included
-        MixedTypeBean nullValues = new MixedTypeBean();
-        assertEquals(aposToQuotes("{'num':null,'plain':null}"),
-                MAPPER.writeValueAsString(nullValues));
-        ObjectMapper mapper;
-
-        // and then change inclusion as property criteria for either
-        mapper = new ObjectMapper();
-        mapper.configOverride(String.class)
-                .setIncludeAsProperty(JsonInclude.Value
-                        .construct(JsonInclude.Include.NON_NULL, null));
-        assertEquals("{\"num\":null}",
-                mapper.writeValueAsString(nullValues));
-
-        mapper = new ObjectMapper();
-        mapper.configOverride(Integer.class)
-                .setIncludeAsProperty(JsonInclude.Value
-                        .construct(JsonInclude.Include.NON_NULL, null));
-        assertEquals("{\"plain\":null}",
-                mapper.writeValueAsString(nullValues));
-    }
-
-    public void testPropConfigOverrideForIncludeAsPropertyAlways() throws Exception
-    {
-        // First, with defaults, only ALWAYS annotated included
-        MixedTypeNonNullBean nullValues = new MixedTypeNonNullBean();
-        assertEquals("{\"annotated\":null}",
-                MAPPER.writeValueAsString(nullValues));
-        ObjectMapper mapper;
-
-        // and then change inclusion as property criteria for either
-        mapper = new ObjectMapper();
-        mapper.configOverride(String.class)
-                .setIncludeAsProperty(JsonInclude.Value
-                        .construct(JsonInclude.Include.ALWAYS, null));
-        assertEquals(aposToQuotes("{'annotated':null,'plain':null}"),
-                mapper.writeValueAsString(nullValues));
-
-        mapper = new ObjectMapper();
-        mapper.configOverride(Integer.class)
-                .setIncludeAsProperty(JsonInclude.Value
-                        .construct(JsonInclude.Include.ALWAYS, null));
-        assertEquals(aposToQuotes("{'num':null,'annotated':null}"),
-                mapper.writeValueAsString(nullValues));
-    }
-
-    public void testConfigOverridesForIncludeAsPropertyNonNull() throws Exception
-    {
-        // First, with ALWAYS override on containing bean, all included
-        MixedTypeNonNullBean nullValues = new MixedTypeNonNullBean();
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.configOverride(MixedTypeNonNullBean.class)
-                .setInclude(JsonInclude.Value
-                        .construct(JsonInclude.Include.ALWAYS, null));
-        assertEquals(aposToQuotes("{'num':null,'annotated':null,'plain':null}"),
-                mapper.writeValueAsString(nullValues));
-
-        // and then change inclusion as property criteria for either
-        mapper = new ObjectMapper();
-        mapper.configOverride(MixedTypeNonNullBean.class)
-                .setInclude(JsonInclude.Value
-                        .construct(JsonInclude.Include.ALWAYS, null));
-        mapper.configOverride(String.class)
-                .setIncludeAsProperty(JsonInclude.Value
-                        .construct(JsonInclude.Include.NON_NULL, null));
-        assertEquals(aposToQuotes("{'num':null,'annotated':null}"),
-                mapper.writeValueAsString(nullValues));
-
-        mapper = new ObjectMapper();
-        mapper.configOverride(MixedTypeNonNullBean.class)
-                .setInclude(JsonInclude.Value
-                        .construct(JsonInclude.Include.ALWAYS, null));
-        mapper.configOverride(Integer.class)
-                .setIncludeAsProperty(JsonInclude.Value
-                        .construct(JsonInclude.Include.NON_NULL, null));
-        assertEquals(aposToQuotes("{'annotated':null,'plain':null}"),
-                mapper.writeValueAsString(nullValues));
-    }
-
-    public void testConfigOverridesForIncludeAsPropertyAlways() throws Exception
-    {
-        // First, with NON_NULL override on containing bean, empty
-        MixedTypeBean nullValues = new MixedTypeBean();
-        ObjectMapper mapper = new ObjectMapper();
-        mapper.configOverride(MixedTypeBean.class)
-                .setInclude(JsonInclude.Value
-                        .construct(JsonInclude.Include.NON_NULL, null));
-        assertEquals("{}",
-                mapper.writeValueAsString(nullValues));
-
-        // and then change inclusion as property criteria for either
-        mapper = new ObjectMapper();
-        mapper.configOverride(MixedTypeBean.class)
-                .setInclude(JsonInclude.Value
-                        .construct(JsonInclude.Include.NON_NULL, null));
-        mapper.configOverride(String.class)
-                .setIncludeAsProperty(JsonInclude.Value
-                        .construct(JsonInclude.Include.ALWAYS, null));
-        assertEquals("{\"plain\":null}",
-                mapper.writeValueAsString(nullValues));
-
-        mapper = new ObjectMapper();
-        mapper.configOverride(MixedTypeBean.class)
-                .setInclude(JsonInclude.Value
-                        .construct(JsonInclude.Include.NON_NULL, null));
-        mapper.configOverride(Integer.class)
-                .setIncludeAsProperty(JsonInclude.Value
-                        .construct(JsonInclude.Include.ALWAYS, null));
-        assertEquals("{\"num\":null}",
-                mapper.writeValueAsString(nullValues));
     }
 
     // [databind#1351], [databind#1417]
